+++ src/main.ts
@@ -1,8 +1,11 @@
 import { createApp } from 'vue'
-import App from './App.vue'
+import { store, key } from './store'
 import router from './router'
+import App from './App.vue'
 import './assets/tailwind.postcss'
 
-createApp(App)
-  .use(router)
-  .mount('#app')
+const app = createApp(App)
+
+app.use(store, key)
+app.use(router)
+app.mount('#app')